Refatorando a aplicação para tranformar o Formulário de Redefinir Minha Senha em um Componente e poder usá-lo ondedesejar, com parâmetros para o texto das validações poder ser diagramável
