+++ src/componentes/EsqueciMinhaSenha/RedefinirSenha/index.js
@@ -1,19 +1,12 @@
 import React, {useState} from "react";
 import {useParams} from 'react-router-dom';
 import {redefinirMinhaSenha} from "../../../services/auth.service";
-import {medidorForcaSenha} from "../../../utils/MedidorForcaSenha";
-import {FormRedefinirSenha} from "../FormRedefinirSenha";
+import {FormRedefinirSenha} from "../../../utils/FormRedefinirSenha";
 
 export const RedefinirSenha = () => {
 
     let {uuid} = useParams();
 
-    const initialValues = {
-        senha: "",
-        confirmacao_senha: "",
-    };
-
-    const [btnOnsubmitReadOnly, setBtnOnsubmitReadOnly] = useState(true);
     const [senhaRedefinida, setSenhaRedefinida] = useState(false);
     const [msgErro, setMsgErro] = useState(false);
 
@@ -24,17 +17,14 @@ export const RedefinirSenha = () => {
             "password2": values.confirmacao_senha
         };
         try {
-            await redefinirMinhaSenha(payload);
+            let texto = await redefinirMinhaSenha(payload);
+            console.log("OnSubmit ", texto)
             setSenhaRedefinida(true);
             setMsgErro(false)
         }catch (e) {
             console.log("Erro ao redefinir senha ", e);
             setMsgErro(true)
         }
-    };
-
-    const validateFormRedefinirSenha = async (values ) => {
-        medidorForcaSenha(values, setBtnOnsubmitReadOnly)
     };
 
     return (
@@ -47,21 +37,11 @@ export const RedefinirSenha = () => {
 
                 <div className='col-12'>
                     <FormRedefinirSenha
-                        initialValues={initialValues}
                         onSubmit={onSubmit}
-                        validateFormRedefinirSenha={validateFormRedefinirSenha}
                         senhaRedefinida={senhaRedefinida}
-                        btnOnsubmitReadOnly={btnOnsubmitReadOnly}
+                        textoValidacaoDentroDoForm={true}
+                        msgErro={msgErro}
                     />
-
-                    {msgErro &&
-                        <div className="alert alert-danger alert-dismissible fade show text-center" role="alert">
-                            Erro ao redefinir a senha, tente novamente
-                            <button type="button" className="close" data-dismiss="alert" aria-label="Close">
-                                <span aria-hidden="true">&times;</span>
-                            </button>
-                        </div>
-                    }
                 </div>
             </div>
         </>
